Mostrando un mensaje de advertencia antes de eliminar de eliminar el producto

+++ src/components/products/Product.js
@@ -3,6 +3,7 @@ import { Link } from 'react-router-dom';
 import Button from '@material-ui/core/Button';
 import DeleteIcon from '@material-ui/icons/Delete';
 import useStyles from './style';
+import Swal from 'sweetalert2'
 
 //Redux
 import { useDispatch } from 'react-redux';
@@ -17,9 +18,22 @@ const Product = ({product}) => {
     //Confirmar si desea eliminar el producto
     const confirmDeleteProduct = id =>{
         //Confirmar
-
-        //pasar dato al action
-        dispatch(deleteProductAction(id));
+        Swal.fire({
+            title: '¿Estas seguro?',
+            text: "Un producto que se elimina no se puede recuperar!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Si, eliminar!',
+            cancelButtonText:'Cancelar'
+          }).then((result) => {
+            if (result.value) {
+                 //pasar dato al action
+                     dispatch(deleteProductAction(id));
+            }
+          })
+       
     }
 
     return(
